+++ frontend/src/pages/nappePhreatique/Nappe_Ph.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import '../../App.css'
 import BackButton from '../../components/BackButton';
+import { Plus } from 'lucide-react';
 
 
 function Nappe_Ph() {
@@ -25,10 +26,24 @@ function Nappe_Ph() {
   };
 
   const handleDelete = async (id) => {
-    await fetch(`http://localhost:8080/api/nappes/${id}`, { method: 'DELETE' });
+  const confirmDelete = window.confirm("Are you sure you want to delete this item?");
+  if (!confirmDelete) return; // Exit if user cancels
+
+  try {
+    const response =     await fetch(`http://localhost:8080/api/nappes/${id}`, { method: 'DELETE' });
+
+    if (response.ok) {
+      alert("Deleted successfully!");
     fetchNappePh();
-  };
-    
+    } else {
+      alert("Failed to delete. Server responded with an error.");
+    }
+  } catch (error) {
+    console.error("Error deleting item:", error);
+    alert("An error occurred while deleting.");
+  }
+};
+
 
   useEffect(() => {
     fetchNappePh();
@@ -40,7 +55,13 @@ function Nappe_Ph() {
         <div className='bg-gray-100'>
       <BackButton />
       <h2 className="text-2xl font-bold mb-4">Management Nappe Phreatique </h2>
-
+   <button
+                onClick={() => navigate('/AddNappeph')}
+                className="flex items-center bg-blue-600 hover:bg-blue-700 text-white font-semibold py-2 px-4 rounded-lg shadow transition duration-200"
+              >
+                <Plus className="w-5 h-5 mr-2" />
+                Ajouter
+              </button>
 
         <table className="w-full bg-white shadow rounded">
           <thead>
